테스트용 임시 단어 파일 분리
